requirements.txt
flask
python-dotenv
streamlit
gradio
langchain
langchain-community
google-generativeai
langchain-google-genai
pypdf2
beautifulsoup4
unstructured
nltk
langchain-groq
faiss-cpu